@@ -1,6 +1,6 @@
 {
 	"name": "execa",
-	"version": "5.1.1",
+	"version": "8.0.1",
 	"description": "Process execution for humans",
 	"license": "MIT",
 	"repository": "sindresorhus/execa",
@@ -10,11 +10,16 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": {
+		"types": "./index.d.ts",
+		"default": "./index.js"
+	},
 	"engines": {
-		"node": ">=10"
+		"node": ">=16.17"
 	},
 	"scripts": {
-		"test": "xo && nyc ava && tsd"
+		"test": "xo && c8 ava && tsd"
 	},
 	"files": [
 		"index.js",
@@ -36,31 +41,33 @@
 		"binaries",
 		"npm",
 		"path",
-		"local"
+		"local",
+		"zx"
 	],
 	"dependencies": {
 		"cross-spawn": "^7.0.3",
-		"get-stream": "^6.0.0",
-		"human-signals": "^2.1.0",
-		"is-stream": "^2.0.0",
+		"get-stream": "^8.0.1",
+		"human-signals": "^5.0.0",
+		"is-stream": "^3.0.0",
 		"merge-stream": "^2.0.0",
-		"npm-run-path": "^4.0.1",
-		"onetime": "^5.1.2",
-		"signal-exit": "^3.0.3",
-		"strip-final-newline": "^2.0.0"
+		"npm-run-path": "^5.1.0",
+		"onetime": "^6.0.0",
+		"signal-exit": "^4.1.0",
+		"strip-final-newline": "^3.0.0"
 	},
 	"devDependencies": {
-		"@types/node": "^14.14.10",
-		"ava": "^2.4.0",
-		"get-node": "^11.0.1",
+		"@types/node": "^20.4.0",
+		"ava": "^5.2.0",
+		"c8": "^8.0.1",
+		"get-node": "^14.2.0",
 		"is-running": "^2.1.0",
-		"nyc": "^15.1.0",
-		"p-event": "^4.2.0",
-		"tempfile": "^3.0.0",
-		"tsd": "^0.13.1",
-		"xo": "^0.35.0"
+		"p-event": "^6.0.0",
+		"path-key": "^4.0.0",
+		"tempfile": "^5.0.0",
+		"tsd": "^0.28.1",
+		"xo": "^0.55.0"
 	},
-	"nyc": {
+	"c8": {
 		"reporter": [
 			"text",
 			"lcov"
@@ -70,5 +77,14 @@
 			"**/test.js",
 			"**/test/**"
 		]
+	},
+	"ava": {
+		"workerThreads": false
+	},
+	"xo": {
+		"rules": {
+			"unicorn/no-empty-file": "off",
+			"@typescript-eslint/ban-types": "off"
+		}
 	}
 }
