@@ -1,7 +1,7 @@
 {
 	"name": "get-stream",
-	"version": "6.0.1",
-	"description": "Get a stream as a string, buffer, or array",
+	"version": "8.0.1",
+	"description": "Get a stream as a string, Buffer, ArrayBuffer or array",
 	"license": "MIT",
 	"repository": "sindresorhus/get-stream",
 	"funding": "https://github.com/sponsors/sindresorhus",
@@ -10,16 +10,21 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": {
+		"types": "./source/index.d.ts",
+		"default": "./source/index.js"
+	},
 	"engines": {
-		"node": ">=10"
+		"node": ">=16"
 	},
 	"scripts": {
-		"test": "xo && ava && tsd"
+		"benchmark": "node benchmarks/index.js",
+		"test": "xo && ava && tsd --typings=source/index.d.ts --files=source/index.test-d.ts"
 	},
 	"files": [
-		"index.js",
-		"index.d.ts",
-		"buffer-stream.js"
+		"source",
+		"!*.test-d.ts"
 	],
 	"keywords": [
 		"get",
@@ -34,14 +39,15 @@
 		"consume",
 		"readable",
 		"readablestream",
-		"array",
-		"object"
+		"object",
+		"concat"
 	],
 	"devDependencies": {
-		"@types/node": "^14.0.27",
-		"ava": "^2.4.0",
-		"into-stream": "^5.0.0",
-		"tsd": "^0.13.1",
-		"xo": "^0.24.0"
+		"@types/node": "^20.5.0",
+		"ava": "^5.3.1",
+		"precise-now": "^2.0.0",
+		"stream-json": "^1.8.0",
+		"tsd": "^0.28.1",
+		"xo": "^0.56.0"
 	}
 }
