@@ -1,18 +1,21 @@
 {
   "name": "human-signals",
-  "version": "2.1.0",
-  "main": "build/src/main.js",
+  "version": "5.0.0",
+  "type": "module",
+  "exports": {
+    "types": "./build/src/main.d.ts",
+    "default": "./build/src/main.js"
+  },
+  "main": "./build/src/main.js",
+  "types": "./build/src/main.d.ts",
   "files": [
-    "build/src",
-    "!~"
+    "build/src/**/*.{js,json,d.ts}",
+    "!build/src/**/*.test.js",
+    "!build/src/{helpers,fixtures}"
   ],
+  "sideEffects": false,
   "scripts": {
     "test": "gulp test"
-  },
-  "husky": {
-    "hooks": {
-      "pre-push": "gulp check --full"
-    }
   },
   "description": "Human-friendly process signals",
   "keywords": [
@@ -20,7 +23,6 @@
     "signals",
     "handlers",
     "error-handling",
-    "errors",
     "interrupts",
     "sigterm",
     "sigint",
@@ -32,33 +34,28 @@
     "operating-system",
     "es6",
     "javascript",
+    "typescript",
     "linux",
     "macos",
     "windows",
     "nodejs"
   ],
   "license": "Apache-2.0",
-  "homepage": "https://git.io/JeluP",
+  "homepage": "https://www.github.com/ehmicky/human-signals",
   "repository": "ehmicky/human-signals",
   "bugs": {
     "url": "https://github.com/ehmicky/human-signals/issues"
   },
   "author": "ehmicky <ehmicky@gmail.com> (https://github.com/ehmicky)",
   "directories": {
-    "lib": "src",
-    "test": "test"
+    "lib": "src"
   },
-  "types": "build/src/main.d.ts",
-  "dependencies": {},
   "devDependencies": {
-    "@ehmicky/dev-tasks": "^0.31.9",
-    "ajv": "^6.12.0",
-    "ava": "^3.5.0",
-    "gulp": "^4.0.2",
-    "husky": "^4.2.3",
-    "test-each": "^2.0.0"
+    "@ehmicky/dev-tasks": "^2.0.80",
+    "ajv": "^8.12.0",
+    "test-each": "^6.0.0"
   },
   "engines": {
-    "node": ">=10.17.0"
+    "node": ">=16.17.0"
   }
 }
