@@ -1,6 +1,6 @@
 {
 	"name": "is-stream",
-	"version": "2.0.1",
+	"version": "3.0.0",
 	"description": "Check if something is a Node.js stream",
 	"license": "MIT",
 	"repository": "sindresorhus/is-stream",
@@ -10,8 +10,10 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=8"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -33,10 +35,10 @@
 		"is"
 	],
 	"devDependencies": {
-		"@types/node": "^11.13.6",
-		"ava": "^1.4.1",
-		"tempy": "^0.3.0",
-		"tsd": "^0.7.2",
-		"xo": "^0.24.0"
+		"@types/node": "^16.4.13",
+		"ava": "^3.15.0",
+		"tempy": "^1.0.1",
+		"tsd": "^0.17.0",
+		"xo": "^0.44.0"
 	}
 }
