@@ -1,16 +1,21 @@
 {
 	"name": "npm-run-path",
-	"version": "4.0.1",
+	"version": "5.2.0",
 	"description": "Get your PATH prepended with locally installed binaries",
 	"license": "MIT",
 	"repository": "sindresorhus/npm-run-path",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
+	"types": "./index.d.ts",
+	"sideEffects": false,
 	"engines": {
-		"node": ">=8"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -34,11 +39,11 @@
 		"executable"
 	],
 	"dependencies": {
-		"path-key": "^3.0.0"
+		"path-key": "^4.0.0"
 	},
 	"devDependencies": {
-		"ava": "^1.4.1",
-		"tsd": "^0.7.2",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"tsd": "^0.17.0",
+		"xo": "^0.45.0"
 	}
 }
