@@ -1,6 +1,6 @@
 {
 	"name": "onetime",
-	"version": "5.1.2",
+	"version": "6.0.0",
 	"description": "Ensure a function is only called once",
 	"license": "MIT",
 	"repository": "sindresorhus/onetime",
@@ -10,8 +10,10 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=6"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -33,11 +35,11 @@
 		"prevent"
 	],
 	"dependencies": {
-		"mimic-fn": "^2.1.0"
+		"mimic-fn": "^4.0.0"
 	},
 	"devDependencies": {
-		"ava": "^1.4.1",
-		"tsd": "^0.7.1",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"tsd": "^0.14.0",
+		"xo": "^0.38.2"
 	}
 }
