@@ -1,6 +1,6 @@
 {
 	"name": "p-limit",
-	"version": "3.1.0",
+	"version": "5.0.0",
 	"description": "Run multiple promise-returning & async functions with limited concurrency",
 	"license": "MIT",
 	"repository": "sindresorhus/p-limit",
@@ -10,15 +10,27 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": {
+		"types": "./index.d.ts",
+		"default": "./index.js"
+	},
+	"imports": {
+		"#async_hooks": {
+			"node": "async_hooks",
+			"default": "./async-hooks-stub.js"
+		}
+	},
 	"engines": {
-		"node": ">=10"
+		"node": ">=18"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
 	},
 	"files": [
 		"index.js",
-		"index.d.ts"
+		"index.d.ts",
+		"async-hooks-stub.js"
 	],
 	"keywords": [
 		"promise",
@@ -38,15 +50,15 @@
 		"bluebird"
 	],
 	"dependencies": {
-		"yocto-queue": "^0.1.0"
+		"yocto-queue": "^1.0.0"
 	},
 	"devDependencies": {
-		"ava": "^2.4.0",
-		"delay": "^4.4.0",
-		"in-range": "^2.0.0",
-		"random-int": "^2.0.1",
-		"time-span": "^4.0.0",
-		"tsd": "^0.13.1",
-		"xo": "^0.35.0"
+		"ava": "^5.3.1",
+		"delay": "^6.0.0",
+		"in-range": "^3.0.0",
+		"random-int": "^3.0.0",
+		"time-span": "^5.1.0",
+		"tsd": "^0.29.0",
+		"xo": "^0.56.0"
 	}
 }
