@@ -1,4 +1,5 @@
-const { LunaDBClient, APIResponse } = require("./index.js");
+import LunaDBClient from "./index.js";
+import { expect, test } from "vitest";
 const nock = require("nock");
 
 test("v0beta: creates document", async () => {
