@@ -1,6 +1,6 @@
 const HEADER_ANALYZE_QUERY = "x-analyze-query";
 
-class LunaDBClient {
+export default class LunaDBClient {
   constructor(db_url) {
     this.db_url = db_url.replace(/\/$/, "");
     this.headers = new Headers();
@@ -70,7 +70,7 @@ class LunaDBClient {
   }
 }
 
-class APIResponse {
+export class APIResponse {
   constructor(status, content) {
     this.status = status;
     this.content = content;
@@ -84,5 +84,3 @@ class APIResponse {
     return this.status >= 200 && this.status < 300;
   }
 }
-
-module.exports = { LunaDBClient, APIResponse };
